Controls now disable correctly in year picker.
Noticed a new bug, though: interestingly, when you select an invalid month (e.g. the first date the calendar allows is in Oct 1987, but we were in Aug 1988 and we go back a year, so that you would end up in the illegal month Aug 1987), it takes you to the ABSOLUTE beginning, whether or not you were past the end or the beginning, and without updating the month.