Month correction logic is working when a month past the start or end dates is selected. Also added keyboard navigation.
Keyboard navigation: added Shift+PageUp + Shift+PageDown to move a year forward or backward, and Ctrl+Shift+PageUp and Ctrl+Shift+PageDown to move 5 years at a time.

However, I think I see why it's easier not to worry about moving days around.  I'm going to leave the always-set-day-of-month-to-1-even-if-not-within-bounds issue for now, since that was the existing logic.  It's a visual display issue, and doesn't seem to cause any invalid days to be selected.